Refactor Overseerr API service into modular extensions
